Refactor code into more files.
